unify the int type for config model

+++ src/bitcoin/reward.go
@@ -4,10 +4,10 @@ const (
 	InitReward = 50
 )
 
-func ComputeReward(lastBlockId uint64, blocksPerRewrad int) uint64 {
+func ComputeReward(lastBlockId uint64, blocksPerRewrad uint64) uint64 {
 	var reward uint64 = InitReward
 	if lastBlockId > 0 {
-		reward /= (lastBlockId+1)/uint64(blocksPerRewrad+1) + 1
+		reward /= (lastBlockId+1)/(blocksPerRewrad+1) + 1
 	}
 	return reward
 }
